+++ release-ci.sh
@@ -72,6 +72,7 @@ echo "Attempting to release $LOCAL_VERSION (old version $BINTRAY_VERSION)"
     :http4k-testing-approval:bintrayUpload \
     :http4k-testing-chaos:bintrayUpload \
     :http4k-testing-hamkrest:bintrayUpload \
+    :http4k-testing-kotest:bintrayUpload \
     :http4k-testing-servirtium:bintrayUpload \
     :http4k-testing-webdriver:bintrayUpload