+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This list is not currently intended to be all-encompassing - it will document ma
 changes with their rationale when appropriate:
 
 ### v3.257.0
-- [http4k-serverless-*] Making the APIs consistent between flavours.
+- [http4k-serverless-*] Making the APIs consistent between flavours by ensu
 
 ### v3.256.1
 - [http4k-core] Fix #470. Path.of cannot decode path parameter values containing %/